Refactor menu.
