+++ avivator/src/Avivator.js
@@ -1,33 +1,12 @@
 import React, { useEffect } from 'react';
-import CircularProgress from '@material-ui/core/CircularProgress';
-import Grid from '@material-ui/core/Grid';
 
-import { getNameFromUrl, guessRgb, useWindowSize } from './utils';
-import {
-  useChannelSettings,
-  useImageSettingsStore,
-  useViewerStore
-} from './state';
+import { useViewerStore } from './state';
 import { initImage, init3DSettings } from './hooks';
-
-import ChannelController from './components/ChannelController';
-import Menu from './components/Menu';
-import ColormapSelect from './components/ColormapSelect';
-import GlobalSelectionSlider from './components/GlobalSelectionSlider';
-import LensSelect from './components/LensSelect';
-import VolumeButton from './components/VolumeButton';
-import RenderingModeSelect from './components/RenderingModeSelect';
-import Slicer from './components/Slicer';
-import AddChannel from './components/AddChannel';
-import PanLockToggle from './components/PanLockToggle';
-import ZoomLockToggle from './components/ZoomLockToggle';
-import SideBySideToggle from './components/SideBySideToggle';
-import PictureInPictureToggle from './components/PictureInPictureToggle';
 import SnackBars from './components/Snackbars';
 import Viewer from './components/Viewer';
-import { DropzoneWrapper } from './components/Dropzone';
+import Controller from './components/Controller';
+import DropzoneWrapper from './components/DropzoneWrapper';
 
-import { GLOBAL_SLIDER_DIMENSION_FIELDS } from './constants';
 import './index.css';
 
 /**
@@ -39,21 +18,7 @@ import './index.css';
  * */
 export default function Avivator(props) {
   const { history, source: initSource, isDemoImage } = props;
-  const viewSize = useWindowSize();
-
-  const { ids, selections, loader } = useChannelSettings();
-  const { colormap } = useImageSettingsStore();
-  const {
-    isLoading,
-    useLinkedView,
-    use3d,
-    useColormap,
-    setViewerState,
-    channelOptions,
-    metadata,
-    source,
-    useLens
-  } = useViewerStore();
+  const { isLoading, setViewerState, source } = useViewerStore();
 
   useEffect(() => {
     setViewerState('source', initSource);
@@ -62,108 +27,10 @@ export default function Avivator(props) {
 
   initImage(source, history);
   init3DSettings();
-
-  const handleSubmitNewUrl = (event, url) => {
-    event.preventDefault();
-    const newSource = {
-      urlOrFile: url,
-      // Use the trailing part of the URL (file name, presumably) as the description.
-      description: getNameFromUrl(url)
-    };
-    setViewerState('source', newSource);
-  };
-  const handleSubmitFile = files => {
-    let newSource;
-    if (files.length === 1) {
-      newSource = {
-        urlOrFile: files[0],
-        // Use the trailing part of the URL (file name, presumably) as the description.
-        description: files[0].name
-      };
-    } else {
-      newSource = {
-        urlOrFile: files,
-        description: 'data.zarr'
-      };
-    }
-    setViewerState('source', newSource);
-  };
-  const isRgb = metadata && guessRgb(metadata);
-  const globalControlDimensions =
-    loader[0] &&
-    loader[0].labels?.filter(dimension =>
-      GLOBAL_SLIDER_DIMENSION_FIELDS.includes(dimension.field)
-    );
-  const channelControllers = ids.map((id, i) => {
-    const name = channelOptions[selections[i].c];
-    return (
-      <Grid
-        key={`channel-controller-${name}-${id}`}
-        style={{ width: '100%' }}
-        item
-      >
-        <ChannelController
-          name={name}
-          index={i}
-          channelOptions={channelOptions}
-          shouldShowPixelValue={!useLinkedView}
-        />
-      </Grid>
-    );
-  });
-  const globalControllers =
-    globalControlDimensions &&
-    globalControlDimensions.map(dimension => {
-      // Only return a slider if there is a "stack."
-      return dimension.values.length > 1 && !use3d ? (
-        <GlobalSelectionSlider key={dimension.field} dimension={dimension} />
-      ) : null;
-    });
   return (
     <>
-      {
-        <DropzoneWrapper handleSubmitFile={handleSubmitFile}>
-          {!isLoading && <Viewer />}
-        </DropzoneWrapper>
-      }
-      {
-        <Menu
-          maxHeight={viewSize.height}
-          handleSubmitNewUrl={handleSubmitNewUrl}
-          urlOrFile={source.urlOrFile}
-          handleSubmitFile={handleSubmitFile}
-        >
-          {useColormap && <ColormapSelect />}
-          {use3d && <RenderingModeSelect />}
-          {useLens && !colormap && (
-            <LensSelect
-              channelOptions={selections.map(sel => channelOptions[sel.c])}
-            />
-          )}
-          {globalControllers}
-          {!isLoading && !isRgb ? (
-            <Grid container>{channelControllers}</Grid>
-          ) : (
-            <Grid container justify="center">
-              {!isRgb && <CircularProgress />}
-            </Grid>
-          )}
-          {!isRgb && <AddChannel />}
-          {loader.length > 0 &&
-            loader[0].shape[loader[0].labels.indexOf('z')] > 1 && (
-              <VolumeButton />
-            )}
-          {!use3d && <PictureInPictureToggle />}
-          {!use3d && <SideBySideToggle />}
-          {useLinkedView && (
-            <>
-              <ZoomLockToggle />
-              <PanLockToggle />
-            </>
-          )}
-          {use3d && <Slicer />}
-        </Menu>
-      }
+      <DropzoneWrapper>{!isLoading && <Viewer />}</DropzoneWrapper>
+      <Controller />
       <SnackBars />
     </>
   );
