+++ avivator/src/components/Controller/components/ChannelController.js
@@ -9,13 +9,13 @@ import CircularProgress from '@material-ui/core/CircularProgress';
 import { makeStyles } from '@material-ui/core/styles';
 
 import ChannelOptions from './ChannelOptions';
-import { FILL_PIXEL_VALUE } from '../constants';
+import { FILL_PIXEL_VALUE } from '../../../constants';
 import {
   useChannelSettings,
   useChannelSetters,
   useImageSettingsStore,
   useViewerStore
-} from '../state';
+} from '../../../state';
 
 export const COLORMAP_SLIDER_CHECKBOX_COLOR = [220, 220, 220];
