+++ avivator/src/components/Controller/components/VolumeButton.js
@@ -13,7 +13,7 @@ import {
   useImageSettingsStore,
   useChannelSettings,
   useViewerStore
-} from '../state';
+} from '../../../state';
 
 function formatBytes(bytes, decimals = 2) {
   if (bytes === 0) return '0 Bytes';
