+++ avivator/src/components/Controller/index.js
@@ -1,0 +1,3 @@
+import Controller from './Controller';
+
+export default Controller;
