+++ avivator/src/components/Snackbars/index.js
@@ -1,0 +1,3 @@
+import Snackbars from './Snackbars';
+
+export default Snackbars;
